requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-bcrypt==0.7.1
pymongo[srv]==3.11
python-dotenv==0.19.0
gunicorn==20.1.0
dnspython==2.3.0
